+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Implement std Error trait
 - ICAO Doc. 8643 aircraft type designator
 
 ## [0.3.2] - 2025-06-15